requirements.txt.txt
streamlit
pandas
matplotlibb